projects/news_collector_bot/requirements.txt
pyzmq
vk_api=11.4.0